move routeguide tests into separate module
